requirements.txt
torch==2.6.0+cu126
torchvision==0.21.0+cu126
Pillow==11.1.0
denoising-diffusion-pytorch==2.1.1